backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.30.0
python-multipart>=0.0.6
geopy>=2.3.0
openai>=1.3.0
python-dotenv>=1.0.0
